+++ tests/testthat/test_checkInteger.R
@@ -29,3 +29,27 @@ test_that("bounds of vectors with only missings are not checked", {
   expect_fail_all(Integer, 0L, lower = 1L)
   expect_fail_all(Integer, 100L, upper = 10L)
 })
+
+test_that("sorted works", {
+  xu = sample(10)
+  while(!is.unsorted(xu))
+    xu = sample(10)
+  xs = sort(xu)
+
+  expect_true(checkInteger(xs, sorted = TRUE))
+  expect_true(grepl("sorted", checkInteger(xu, sorted = TRUE), fixed = TRUE))
+
+  expect_true(checkInteger(1L, sorted = TRUE))
+  expect_true(checkInteger(integer(0), sorted = TRUE))
+  expect_true(checkInteger(NA_integer_, sorted = TRUE))
+  expect_true(checkInteger(rep(NA_integer_, 10), sorted = TRUE))
+
+  for (i in 1:10) {
+    x = sample(10)
+    x[sample(10, sample(7:9, 1))] = NA
+    if (is.unsorted(na.omit(x)))
+      expect_true(grepl("sorted", checkInteger(xu, sorted = TRUE), fixed = TRUE))
+    else
+      expect_true(grepl("sorted", checkInteger(xu, sorted = TRUE), fixed = TRUE))
+  }
+})